backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
itsdangerous==2.1.2
httpx==0.26.0
resend==0.7.0
python-multipart==0.0.6
email-validator==2.1.0
